+++ benchmarks/metadata_extraction/dataclass.py
@@ -1,16 +1,13 @@
-from typing import List, Optional, Dict
-from dataclasses import dataclass, field
+from typing import List, Optional
 from pydantic import BaseModel
 
 
-@dataclass
-class Metadata:
+class Metadata(BaseModel):
     send_date: Optional[List[str]]
     letter_title: Optional[List[str]]
     sender_persons: Optional[List[str]]
     receiver_persons: Optional[List[str]]
 
 
-@dataclass
 class Document(BaseModel):
     metadata: Metadata